Adds user-created workouts

+++ components/EditButton/EditButton.jsx
@@ -1,5 +1,3 @@
-// istanbul ignore file
-
 // React
 import React, { useState } from 'react';
 
@@ -13,13 +11,88 @@ import DropdownButton from 'react-bootstrap/DropdownButton';
 import Modal from 'react-bootstrap/Modal';
 
 // Firebase
-import { deleteDoc, doc } from 'firebase/firestore';
+import { deleteDoc, doc, updateDoc } from 'firebase/firestore';
 import { db } from '../../firebase-config';
 
 // Custom components
 import CustomToast from './CustomToast';
 
+// Authentication
+import { useAuth } from '../../context/authUserContext';
+
+// Makes either a button, or a dropdown button
+function AButton({ type, name, id, handleModalOpen }) {
+  if (type === 'user') {
+    return (
+      <DropdownButton title="...">
+        <Link href={`/userworkouts/edit/${id}`} passHref>
+          <Dropdown.Item>Edit {name}</Dropdown.Item>
+        </Link>
+        <Dropdown.Item onClick={handleModalOpen}>Delete {name}</Dropdown.Item>
+      </DropdownButton>
+    );
+  }
+
+  if (type === 'exercise') {
+    return (
+      <DropdownButton title="...">
+        <Link href={`/exercises/edit/${id}`} passHref>
+          <Dropdown.Item>Edit {name}</Dropdown.Item>
+        </Link>
+        <Dropdown.Item onClick={handleModalOpen}>Delete {name}</Dropdown.Item>
+      </DropdownButton>
+    );
+  }
+
+  if (type === 'workout') {
+    return (
+      <DropdownButton title="...">
+        <Link href={`/workouts/edit/${id}`} passHref>
+          <Dropdown.Item>Edit {name}</Dropdown.Item>
+        </Link>
+        <Dropdown.Item onClick={handleModalOpen}>Delete {name}</Dropdown.Item>
+      </DropdownButton>
+    );
+  }
+}
+
+function DeleteModal({
+  name,
+  type,
+  isModalOpen,
+  handleModalClose,
+  handleDelete,
+}) {
+  return (
+    <Modal show={isModalOpen} onHide={handleModalClose} centered size="lg">
+      <Modal.Header>
+        <Modal.Title>
+          {type === 'user' ? (
+            <p>Delete your workout &#39;{name}&#39;?</p>
+          ) : (
+            <p>
+              Delete {type} &#39;{name}&#39;?
+            </p>
+          )}
+        </Modal.Title>
+      </Modal.Header>
+
+      <Modal.Footer>
+        <Button variant="secondary" onClick={handleModalClose}>
+          Close
+        </Button>
+
+        <Button variant="danger" onClick={handleDelete}>
+          Delete
+        </Button>
+      </Modal.Footer>
+    </Modal>
+  );
+}
+
 export default function EditButton({ type, id, name, onDelete }) {
+  const { authUser } = useAuth();
+
   /* Handles state for the delete toast */
   const [isToastActive, setToastActive] = useState({});
   const handleToastOpen = ({ title, body, error }) => {
@@ -37,6 +110,16 @@ export default function EditButton({ type, id, name, onDelete }) {
   };
 
   const handleDelete = () => {
+    if (type === 'user') {
+      const filtered = authUser.createdWorkouts.filter(
+        (workout) => workout.id !== id
+      );
+      authUser.createdWorkouts = filtered;
+      updateDoc(doc(db, 'users', authUser.uid), {
+        createdWorkouts: filtered,
+      });
+    }
+
     if (type === 'exercise') {
       deleteDoc(doc(db, 'exercises', id))
         .then(() => {
@@ -51,7 +134,9 @@ export default function EditButton({ type, id, name, onDelete }) {
             error,
           });
         });
-    } else {
+    }
+
+    if (type === 'workout') {
       deleteDoc(doc(db, 'workouts', id))
         .then(() => {
           handleToastOpen({
@@ -70,51 +155,6 @@ export default function EditButton({ type, id, name, onDelete }) {
     handleModalClose();
   };
 
-  // Makes either a button, or a dropdown button
-  const makeButton = (title) => (
-    <DropdownButton title="...">
-      {title === 'exercise' ? (
-        <Link href={`/exercises/edit/${id}`} passHref>
-          <Dropdown.Item>Edit {title}</Dropdown.Item>
-        </Link>
-      ) : (
-        <Link href={`/workouts/edit/${id}`} passHref>
-          <Dropdown.Item>Edit {title}</Dropdown.Item>
-        </Link>
-      )}
-
-      <Dropdown.Item onClick={handleModalOpen}>Delete {title}</Dropdown.Item>
-    </DropdownButton>
-  );
-
-  // Creates the modal only if there is an id.
-  const makeModal = (toShow) => {
-    if (toShow !== undefined) {
-      return (
-        <Modal show={isModalOpen} onHide={handleModalClose} centered size="lg">
-          <Modal.Header>
-            <Modal.Title>
-              <p>
-                Delete {type} &#39;{name}&#39;?
-              </p>
-            </Modal.Title>
-          </Modal.Header>
-
-          <Modal.Footer>
-            <Button variant="secondary" onClick={handleModalClose}>
-              Close
-            </Button>
-
-            <Button variant="danger" onClick={handleDelete}>
-              Delete
-            </Button>
-          </Modal.Footer>
-        </Modal>
-      );
-    }
-    return null;
-  };
-
   // Creates the toast for deleting
   function makeToast({ title, body, error }) {
     if (title !== undefined) {
@@ -129,11 +169,23 @@ export default function EditButton({ type, id, name, onDelete }) {
     }
     return null;
   }
-
   return (
     <>
-      {makeButton(type)}
-      {makeModal(id)}
+      <AButton
+        type={type}
+        name={name}
+        id={id}
+        handleModalOpen={handleModalOpen}
+      />
+      {id !== undefined && (
+        <DeleteModal
+          name={name}
+          type={type}
+          isModalOpen={isModalOpen}
+          handleModalClose={handleModalClose}
+          handleDelete={handleDelete}
+        />
+      )}
       {makeToast(isToastActive)}
     </>
   );
